requirements.txt
fastmcp
python-pptx
python-docx